move everything related to config-variables to a dedicated folder
