+++ packages/twenty-front/src/modules/settings/admin-panel/config-variables/graphql/queries/getConfigVariablesGrouped.ts
